requirements.txt
Django>=5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
gunicorn>=20.1
drf-spectacular>=0.27
djangorestframework-simplejwt>=5.3
django-filter>=22.1
